+++ helpers.h
@@ -24,13 +24,13 @@ void testDatasets(std::vector<Data> &trainSet, std::vector<Data> &valSet, std::v
 void saveHistory(History history, const char* fileName);
 int cudaDeviceProperties();
 
-void hostBatchPreds(float* output_activations, int * batch_pred, int output_size, int b_size);
+void hostBatchPreds(float* output_activations, unsigned short * batch_pred, int output_size, int b_size);
 void hostElementMult(float *h_M, float *h_N, float *h_P, int num_MRows, int num_MCols, int num_NRows, int num_NCols);
 void hostDotProduct(float* M, float* N, float* P, int num_MRows, int num_MCols, int num_NRows, int num_NCols);
 void hostActivationFuncForward(float *h_Z, float *h_Y, int numRows, int numCols);
 void hostActivationFuncBackward(float *h_Z, float *h_dervA, float *h_dervZ, int numRows, int numCols);
 void hostTranspose(float *h_M, float *h_N, int num_MRows, int num_MCols);
-float hostMSE(float* h_T, float* h_O, int batchSize, int numLabels);
+float hostMSE(unsigned short* h_T, float* h_O, int batchSize, int numLabels);
 
 
 #endif // HELPERS_H
